+++ packages/code-du-travail-data/src/prime-precarite/PrimePrecarite.ts
@@ -1,0 +1,4 @@
+import Data from "./precarite.data.json";
+import type { PrecariteSituation } from "./types";
+
+export const primePrecariteData: PrecariteSituation[] = Data;